requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-filter
django-rest-passwordreset
djangorestframework
djangorestframework-simplejwt
filelock
fontawesomefree
gunicorn
Markdown
packaging
pillow
platformdirs
psycopg2-binary
PyJWT
sqlparse
typing-extensions
tzdata
virtualenv
whitenoise
requests
python-dotenv

resend
cloudinary
weasyprint
xhtml2pdf

# playwright
beautifulsoup4
lxml